requirements.txt
flask==2.0.1
python-docx==0.8.11
Pillow==9.5.0
gunicorn==20.1.0
